+++ ThirdPartyAdapters/adcolony/adcolony/src/main/java/com/google/ads/mediation/adcolony/AdColonyBannerRenderer.java
@@ -32,8 +32,8 @@ public class AdColonyBannerRenderer extends AdColonyAdViewListener implements Me
   private final MediationBannerAdConfiguration adConfiguration;
 
   public AdColonyBannerRenderer(
-          @NonNull MediationBannerAdConfiguration adConfiguration,
-          @NonNull MediationAdLoadCallback<MediationBannerAd, MediationBannerAdCallback> callback
+      @NonNull MediationBannerAdConfiguration adConfiguration,
+      @NonNull MediationAdLoadCallback<MediationBannerAd, MediationBannerAdCallback> callback
   ) {
     this.adLoadCallback = callback;
     this.adConfiguration = adConfiguration;
@@ -42,26 +42,24 @@ public class AdColonyBannerRenderer extends AdColonyAdViewListener implements Me
   public void render() {
     if (adConfiguration.getAdSize() == null) {
       AdError error = createAdapterError(ERROR_INVALID_SERVER_PARAMETERS,
-              "Failed to request banner with unsupported size: null");
+          "Failed to request banner with unsupported size: null");
       Log.e(TAG, error.getMessage());
       adLoadCallback.onFailure(error);
       return;
     }
 
-    AdColonyAdOptions adOptions = AdColonyManager
-            .getInstance()
-            .getAdOptionsFromAdConfig(adConfiguration);
-    ArrayList<String> listFromServerParams = AdColonyManager
-            .getInstance()
-            .parseZoneList(adConfiguration.getServerParameters());
-    String requestedZone = AdColonyManager
-            .getInstance()
-            .getZoneFromRequest(listFromServerParams, adConfiguration.getMediationExtras());
+    AdColony.setAppOptions(AdColonyManager.getInstance().buildAppOptions(adConfiguration));
+    AdColonyAdOptions adOptions = AdColonyManager.getInstance()
+        .getAdOptionsFromAdConfig(adConfiguration);
+    ArrayList<String> listFromServerParams = AdColonyManager.getInstance()
+        .parseZoneList(adConfiguration.getServerParameters());
+    String requestedZone = AdColonyManager.getInstance()
+        .getZoneFromRequest(listFromServerParams, adConfiguration.getMediationExtras());
     // Setting the requested size as the AdColony view size
-    AdColonyAdSize adSize = new AdColonyAdSize(
-            convertPixelsToDp(adConfiguration.getAdSize().getWidthInPixels(adConfiguration.getContext())),
-            convertPixelsToDp(adConfiguration.getAdSize().getHeightInPixels(adConfiguration.getContext()))
-    );
+    AdColonyAdSize adSize = new AdColonyAdSize(convertPixelsToDp(
+        adConfiguration.getAdSize().getWidthInPixels(adConfiguration.getContext())),
+        convertPixelsToDp(
+            adConfiguration.getAdSize().getHeightInPixels(adConfiguration.getContext())));
     AdColony.requestAdView(requestedZone, this, adSize, adOptions);
   }